KON-63: Fixed reminders
