+++ Resources/views/main/index.html.twig
@@ -22,6 +22,8 @@
 
       <a class="btn btn-info btn-sm w-auto" href="{{ path('process_new') }}" role="button">{{ 'navigation.main.new_process'|trans }}</a>
 
+      <a class="btn btn-warning btn-sm" href="{{ path('user_reminder_index') }}" role="button">{{ 'navigation.main.remainders.number_of_reminders'|trans({'%numberOfReminders%': activeUserReminders|length}) }}</a>
+
       <ul class="navbar-nav px-3 w-auto">
         <li class="nav-item text-nowrap">
           <a class="nav-link" href="{{ path('fos_user_security_logout') }}">{{ 'navigation.main.sign_out'|trans }}</a>
@@ -34,6 +36,7 @@
     <div class="row">
       <nav class="col-md-2 d-none d-md-block bg-light sidebar">
         <div class="sidebar-sticky">
+          <h3>Main menu</h3>
           {% block navigation_sidebar %}
           <ul class="nav flex-column">
             {% for menuItem in menuItems %}
@@ -46,7 +49,8 @@
             {% endfor %}
           </ul>
 
-            SUBMENU
+
+            <h3>Sub menu</h3>
             <ul class="nav flex-column">
               {% for menuItem in subMenuItems %}
                 <li class="nav-item">
