+++ Resources/views/reminder/index.html.twig
@@ -3,37 +3,39 @@
 {% block title %}{{ 'reminder.index.title'|trans }}{% endblock %}
 
 {% block content %}
-    <h1>{{ 'reminder.index.title'|trans }}</h1>
+  <h1>{{ 'reminder.index.title'|trans }}</h1>
 
-    <table class="table">
-        <thead>
-            <tr>
-                <th>{{ 'reminder.table.id'|trans }}</th>
-                <th>{{ 'reminder.table.message'|trans }}</th>
-                <th>{{ 'reminder.table.date'|trans }}</th>
-                <th>{{ 'reminder.table.finished'|trans }}</th>
-                <th>{{ 'common.table.actions'|trans }}</th>
-            </tr>
-        </thead>
-        <tbody>
-        {% for reminder in reminders %}
-            <tr>
-                <td>{{ reminder.id }}</td>
-                <td>{{ reminder.message }}</td>
-                <td>{{ reminder.date ? reminder.date|date('Y-m-d H:i:s') : '' }}</td>
-                <td>{{ (reminder.finished ? 'common.boolean.true' : 'common.boolean.false')|trans }}</td>
-                <td>
-                    <a href="{{ path('reminder_show', {'id': reminder.id, 'process': process.id}) }}">show</a>
-                    <a href="{{ path('reminder_edit', {'id': reminder.id, 'process': process.id}) }}">edit</a>
-                </td>
-            </tr>
-        {% else %}
-            <tr>
-                <td colspan="4">{{ 'common.no_records_found'|trans }}</td>
-            </tr>
-        {% endfor %}
-        </tbody>
-    </table>
+  <table class="table">
+    <thead>
+    <tr>
+      <th>{{ 'reminder.table.id'|trans }}</th>
+      <th>{{ 'reminder.table.message'|trans }}</th>
+      <th>{{ 'reminder.table.date'|trans }}</th>
+      <th>{{ 'reminder.table.finished'|trans }}</th>
+      <th>{{ 'common.table.actions'|trans }}</th>
+    </tr>
+    </thead>
+    <tbody>
+    {% for reminder in reminders %}
+      <tr>
+        <td>{{ reminder.id }}</td>
+        <td>{{ reminder.message }}</td>
+        <td>{{ reminder.date ? reminder.date|date('Y-m-d H:i:s') : '' }}</td>
+        <td>{{ (reminder.finished ? 'common.boolean.true' : 'common.boolean.false')|trans }}</td>
+        <td>
+          <a href="{{ path('reminder_show', {'id': reminder.id, 'process': reminder.process.id}) }}">show</a>
+          <a href="{{ path('reminder_edit', {'id': reminder.id, 'process': reminder.process.id}) }}">edit</a>
+        </td>
+      </tr>
+    {% else %}
+      <tr>
+        <td colspan="4">{{ 'common.no_records_found'|trans }}</td>
+      </tr>
+    {% endfor %}
+    </tbody>
+  </table>
 
+  {% if process is defined %}
     <a href="{{ path('reminder_new', {'process': process.id}) }}">{{ 'common.actions.create'|trans }}</a>
+  {% endif %}
 {% endblock %}
